Change package name from `repositories` to `repository`.
