+++ api-java/src/main/java/org/spine3/examples/todolist/repository/RepositoryHelper.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.repositories;
+package org.spine3.examples.todolist.repository;
 
 import com.google.protobuf.Message;
 import org.spine3.examples.todolist.TaskListId;
