+++ api-java/src/main/java/org/spine3/examples/todolist/repository/TaskLabelsRepository.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.repositories;
+package org.spine3.examples.todolist.repository;
 
 import org.spine3.examples.todolist.TaskId;
 import org.spine3.examples.todolist.c.aggregate.TaskLabelsPart;
